Refactor: Rename and rearrange functions
To make it more clear, which functions just add data about apps. And to
collect functions which operate on multiple apps more in `apps.ts`.
