+++ src/main/appInfo.ts
@@ -10,7 +10,6 @@ import type { AppInfo, PackageJson } from 'pc-nrfconnect-shared';
 
 import {
     AppSpec,
-    AppWithError,
     DownloadableApp,
     InstalledDownloadableApp,
     LocalApp,
@@ -22,7 +21,7 @@ import { getAppsLocalDir, getAppsRootDir, getNodeModulesDir } from './config';
 import describeError from './describeError';
 import { createJsonFile, ifExists, readJsonFile } from './fileUtil';
 import { downloadToFile, downloadToJson } from './net';
-import { downloadAllSources, getAppUrls, getSource } from './sources';
+import { getAppUrls, getSource } from './sources';
 
 export const installedAppPath = (app: AppSpec) => {
     const appDir =
@@ -120,23 +119,17 @@ const downloadIconAndReleaseNotes = (appInfo: AppInfo, source: SourceName) => {
     ]);
 };
 
-export const createDownloadableApp =
+export const addDownloadAppData =
     (source: SourceName) =>
     (appInfo: AppInfo): DownloadableApp => ({
-        name: appInfo.name,
+        ...appInfo,
+
         source,
-
-        displayName: appInfo.displayName,
-        description: appInfo.description,
-
         iconPath: iconPath({ name: appInfo.name, source }),
         releaseNotes: readReleaseNotes({
             name: appInfo.name,
             source,
         }),
-        homepage: appInfo.homepage,
-        latestVersion: appInfo.latestVersion,
-        versions: appInfo.versions,
     });
 
 export const downloadAppInfos = async (source: Source) => {
@@ -166,7 +159,7 @@ export const downloadAppInfos = async (source: Source) => {
     return downloadableApps.filter(defined);
 };
 
-export const getInstalledApp = (
+export const addInstalledAppData = (
     app: DownloadableApp
 ): InstalledDownloadableApp | WithdrawnApp => {
     const appPath = installedAppPath(app);
@@ -177,8 +170,9 @@ export const getInstalledApp = (
     );
 
     return {
+        ...app,
+
         name: packageJson.name,
-        source: app.source,
 
         description: packageJson.description ?? app.description,
         displayName: packageJson.displayName ?? app.displayName,
@@ -186,8 +180,6 @@ export const getInstalledApp = (
         engineVersion: packageJson.engines?.nrfconnect,
 
         currentVersion: packageJson.version,
-        latestVersion: app.latestVersion,
-        versions: app.versions,
 
         path: appPath,
         iconPath:
@@ -196,13 +188,11 @@ export const getInstalledApp = (
 
         homepage: packageJson.homepage ?? app.homepage,
         repositoryUrl: packageJson.repository?.url,
-
-        releaseNotes: app.releaseNotes,
     };
 };
 
 export const getLocalApp = (appName: string): LocalApp => ({
-    ...getInstalledApp({
+    ...addInstalledAppData({
         name: appName,
         source: LOCAL,
         displayName: appName,
@@ -213,64 +203,3 @@ export const getLocalApp = (appName: string): LocalApp => ({
     }),
     source: LOCAL,
 });
-
-export const addInformationForInstalledApps = (
-    downloadableApps: DownloadableApp[]
-) => {
-    const appsWithErrors: AppWithError[] = [];
-    const apps: DownloadableApp[] = [];
-
-    downloadableApps.forEach(app => {
-        if (!isInstalled(app)) {
-            apps.push(app);
-            return;
-        }
-
-        try {
-            apps.push(getInstalledApp(app));
-        } catch (error) {
-            appsWithErrors.push({
-                reason: error,
-                path: installedAppPath(app),
-                name: app.name,
-                source: app.source,
-            });
-        }
-    });
-
-    return { apps, appsWithErrors };
-};
-
-export const downloadLatestAppInfos = async () => {
-    const { sources, sourcesWithErrors } = await downloadAllSources();
-
-    const downloadableAppsPromises = sources.map(async source =>
-        (await downloadAppInfos(source)).map(createDownloadableApp(source.name))
-    );
-    const downloadableApps = (
-        await Promise.all(downloadableAppsPromises)
-    ).flat();
-
-    const { apps, appsWithErrors } =
-        addInformationForInstalledApps(downloadableApps);
-
-    return {
-        apps,
-        appsWithErrors,
-        sourcesWithErrors,
-    };
-};
-
-const getAllAppSpecs = (source: Source): AppSpec[] => {
-    const filesToExclude = ['source.json', 'apps.json', 'updates.json'];
-
-    return fs
-        .readdirSync(getAppsRootDir(source.name))
-        .filter(name => !filesToExclude.includes(name))
-        .filter(name => name.endsWith('.json'))
-        .map(name => name.replace(/\.json$/, ''))
-        .map(name => ({ name, source: source.name }));
-};
-
-export const readAppInfos = (source: Source) =>
-    getAllAppSpecs(source).map(readAppInfo);
